Fixes broken body dump, reauth, backoff and retry
The previous patch to aiclib introduced a bug where the body wouldn't
get dumped to JSON correctly. Additionally, if at any point the request
failed due to authentication, the command would successfully
re-authenticate but still return the failure, instead of retrying. This
patch ameliorates these issues. Finally, backoff was being set to None
and and retry was busted because the body wasn't being reassigned in
the open_kwargs on retry